Add files via upload
Plan Pilot 0.0.1 Final Version
Plan Pilot is a Google Chrome extension created as an open-source project that aims to assist users in managing their to-do lists with ease. This is the 0.0.1 Final Version of the extension, which includes several changes and improvements over previous versions.

Additional Changes
Added the ability to create multiple lists based on user needs
Improved the user interface with a clean, minimalist design that is easy on the eyes
Updated the codebase to improve the extension's performance and stability
Fixed various bugs and issues reported by users
Differences from Version 2
The Final Version includes several bug fixes that were not present in Version 2. Such as tasks that were not being inputted at all.
The Final Version features an updated user interface with a cleaner design and logo
The Final Version includes the ability to create multiple lists based on user needs, which was not available in Version 2

Differences from Version 1
The Final Version includes several improvements to the extension's functionality and user interface that were not present in Version 1
The Final Version features a cleaner, more minimalist design than Version 1
The Final Version includes the ability to create multiple lists based on user needs, which was not available in Version 1. As Version 1 served as a template for future versions.

About Plan Pilot
Plan Pilot is designed to be a simple, user-friendly extension that helps users stay organized and focused. With basic functionality that allows users to create and manage to-do lists with minimal effort, Plan Pilot is perfect for anyone looking to stay on top of their tasks without the need for complicated project management software or apps.

The extension's codebase is well-organized and easy to understand, making it accessible for developers who want to build upon or modify the codebase. As an open-source project, Plan Pilot has the potential to be improved upon and expanded by the developer community.

Overall, Plan Pilot is a useful tool for anyone looking to manage their to-do lists efficiently. The extension's simplicity and ease of use make it a valuable addition to anyone's productivity toolkit. With the potential for future updates and improvements, the extension is poised to become an even more useful tool for users in the future.